requirements.txt
Django==4.2.7
Pillow>=10.0.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
django-allauth>=0.57.0
django-widget-tweaks>=1.4.12
python-decouple>=3.8
django-extensions>=3.2.3
python-telegram-bot>=21.7
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
pandas>=2.2.0
openpyxl>=3.1.5
numpy>=1.26.4
dj-database-url>=2.1.0
requests>=2.31.0